requirements.txt
asarPy==1.0.1
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.3
djangorestframework==3.14.0
jwt==1.3.1
psycopg==3.1.18
psycopg2-binary==2.9.9
py3dns==4.0.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
validate_email==1.3
